test(button): migrate size
