+++ core/src/components/button/test/size/button.e2e.ts-snapshots/button-size-large-in-buttons-ios-ltr-Mobile-Chrome-linux.png
