+++ core/src/components/button/test/size/button.e2e.ts-snapshots/button-size-large-md-ltr-Mobile-Firefox-linux.png
