+++ core/src/components/button/test/size/button.e2e.ts-snapshots/button-size-small-md-ltr-Mobile-Firefox-linux.png
